requirements.txt
asgiref==3.4.1
attrs==21.2.0
Automat==20.2.0
autopep8==1.6.0
backports.entry-points-selectable==1.1.0
certifi==2021.10.8
charset-normalizer==2.0.7
constantly==15.1.0
distlib==0.3.3
Django==3.2.9
django-cors-headers==3.10.0
djangorestframework==3.12.4
filelock==3.2.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
mysql-connector-python==8.0.26
numpy==1.21.2
pipenv==2021.11.15
platformdirs==2.4.0
protobuf==3.18.1
pycodestyle==2.8.0
pymodbus==2.5.2
pyserial==3.5
pytz==2021.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
Twisted==21.7.0
typing-extensions==3.10.0.2
urllib3==1.26.7
virtualenv==20.8.1
virtualenv-clone==0.5.7
zope.interface==5.4.0
